+++ docs/Whats-New.md
@@ -101,6 +101,7 @@ Vanilla fixes:
 - Fix a bug where carryalls would allow boarding in some situations when they shouldn't (by ZivDero)
 - Fix a bug where under some circumstances Tiberium would halt spreading and/or growing temporarily (by ZivDero)
 - Fix a bug where Tiberium spawned by animations wouldn't grow or spread (by ZivDero)
+- Fix a bug where carryalls assign their ROT to the unit they're carrying (by ZivDero)
 
 :::